+++ packages/jest-environment/jest.config.ts
@@ -1,4 +1,4 @@
-import buildConfig from '../../jest.config'
+import buildConfig from '../edge-runtime/jest.config'
 import type { Config } from '@jest/types'
 
 const config: Config.InitialOptions = buildConfig(__dirname)