+++ packages/runtime/src/server/run-server.ts
@@ -1,5 +1,5 @@
 import { createHandler, Options } from './create-handler'
-import type { EdgeContext } from '@edge-runtime/vm'
+import type { EdgeContext } from '@runtime-edge/vm'
 import listen from 'async-listen'
 import http from 'http'
 import type { ListenOptions } from 'net'